WIP - Move setup/teardown logic for vulkan instance and device
